+++ src/colour-from-r.h
@@ -1,2 +1,0 @@
-
-uint32_t rcolour_to_int(const char *rcol);